ARM: dts: aspeed: g6: Remove unused aspeed,ast2600-lhc node
For the AST2500 the node was (previously) used for pinctrl purposes,
and while the hardware capability is also present in the AST2600, the
AST2600 pinctrl no relationship to it. Further, there's no corresponding
binding, so as for the AST2400 and AST2500, remove the node for now to
eliminate the warnings.

Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>